+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   [HexaPDF::Document::Annotations#create_rectangle]
 * [HexaPDF::Type::Annotations::Circle] for ellipse annotations as well as
   [HexaPDF::Document::Annotations#create_ellipse]
+* Basic appearance generation for push button fields
 * [HexaPDF::Type::Annotation::BorderEffect] type class
 * [HexaPDF::Type::Annotations::BorderEffect] module that provides convenience
   access to the border effect dictionary